requirements.txt
asn1crypto==1.5.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docker==7.0.0
exceptiongroup==1.2.0
Flask==3.0.0
greenlet==3.0.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mirakuru==2.5.2
packaging==23.2
pg8000==1.30.3
pgmock==1.3.7
pluggy==1.3.0
port-for==0.7.2
psutil==5.9.6
psycopg==3.1.14
psycopg2==2.9.9
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.23
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
